Unbreak task lock
Deleting stale locks would break because after self.session.delete(lock),
the lock is still in the session as long as it has not been committed.
Instead of committing, and risking a concurrency issue, reuse the lock by
updating its pid and updateDate fields.

Made the code a bit easier to test by adding named parameters for pid
and date, added a unit test.