Improve handling of merge failures
Let the user retry or abort a failed merge.

With this commit, pull.pull does not try to handle the merge it raises an
exception, which is not handled by anyone. Since this happens while
syncmanager is in a "mergeOperation", the BEFORE_MERGE_TAG is not removed.

Next time one tries to use a Yokadi command, the merge check fails and the
user can decide if he wants to retry (in case another Yokadi process is
running) or abort the merge, which means resetting the repo to the state
at BEFORE_MERGE_TAG and deleting the tag.